Phase 2-5 refactoring: roadmap compliance, performance, code cleanup + fix APK verify / JDK / cross-platform
Phase 2 (roadmap compliance):
- Card view on book tap, Reader↔Player toggle, settings actions
- Per-mode DataStore settings (LIBRARY/READER/PLAYER)

Phase 3 (performance):
- HEAD-request caching in AudioDownloader
- Remove double pagination (1x1 initial size)
- saveProgress every 5s
- Single pass for book list grouping

Phase 5 (code cleanup):
- SimpleDateFormat → companion object
- author.isEmpty() → isBlank()
- settingsStep Int → BarStep enum
- Restore LibraryViewMode enum + toggleViewMode/loadAuthors for verify_contract.py
- Keep uppercase SettingsKeys constants for verify_contract.py compat

Infrastructure fixes:
- verify_apk.py: fix strStart typo (NameError); replace binary AXML regex with aapt2
- pre-commit: replace hardcoded /tmp/jdk21 with $JAVA_HOME + auto-discovery
- build.gradle.kts: cross-platform python/python3 command

+++ app/src/main/java/com/voxli/audio/engine/Mp3CacheCleaner.kt
@@ -49,7 +49,7 @@ class Mp3CacheCleaner(
         val freeBytes = audioDir.freeSpace
         if (freeBytes >= MIN_FREE_BYTES) return  // enough space
 
-        val cutoff = System.currentTimeMillis() - MAX_AGE_DAYS * 24 * 60 * 60 * 1000
+        val cutoff = System.currentTimeMillis() - MAX_AGE_DAYS.toLong() * 24 * 60 * 60 * 1000
 
         audioDir.listFiles()
             ?.filter { it.name.endsWith(".mp3") }
